% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{adjust_subtype}
\alias{adjust_subtype}
\title{For subtype permissions: use reference permission to identify R3 & lapse}
\usage{
adjust_subtype(df_subtype, ref_name, db_history)
}
\arguments{
\item{df_subtype}{subtype license history table}

\item{ref_name}{name of permission that provides R3 & lapse (for subtypes)}

\item{db_history}{file path to history.sqlite3}
}
\description{
For subtype permissions: use reference permission to identify R3 & lapse
}
\seealso{
Other functions for producing license history: \code{\link{build_history}},
  \code{\link{load_all}}, \code{\link{write_history}}
}
\concept{functions for producing license history}
