% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{aggregate_pop}
\alias{aggregate_pop}
\title{Convenience function: aggregate population data to 7 age categories}
\usage{
aggregate_pop(pop)
}
\arguments{
\item{pop}{data frame: input population data for state}
}
\description{
Collapses the larger number of age categories provided by acs
}
\seealso{
Other functions to prepare data for summarization: \code{\link{extrapolate_pop}},
  \code{\link{factor_age_acs}}, \code{\link{prep_pop}},
  \code{\link{quarterly_filter}},
  \code{\link{quarterly_lapse}},
  \code{\link{recode_history}}
}
\concept{functions to prepare data for summarization}
