% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{build_history}
\alias{build_history}
\title{Build history table for given permission}
\usage{
build_history(sale, lic, yrs, lic_filter, quarter,
  rank_var = c("duration", "res"), carry_vars = c("month", "res"))
}
\arguments{
\item{sale}{input sales}

\item{lic}{input license types}

\item{yrs}{years to include in license history}

\item{lic_filter}{query to be passed to filter_() on lic table (selects relevant lic_ids)}

\item{quarter}{current quarter (if not quarter 4, the last year is excluded for lapse)}

\item{rank_var}{passed to rank_sale()}

\item{carry_vars}{passed to make_history()}
}
\description{
Build history table for given permission
}
\seealso{
Other functions for producing license history: \code{\link{adjust_subtype}},
  \code{\link{load_all}}, \code{\link{write_history}}
}
\concept{functions for producing license history}
