% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate all 5 metrics for a permission}
\usage{
calc_metrics(history, pop_county, sale, dashboard_yrs, part_ref = NULL,
  res_type = NULL, segs = c("tot", "res", "sex", "agecat", "county"),
  tests = c(tot = 20, res = 35, sex = 35, agecat = 35, county = 70),
  tests_recruits = c(tot = 30, res = 45, sex = 45, agecat = 45, county =
  150), scaleup_test = 25, rate_test = 1)
}
\arguments{
\item{history}{data frame: input history table for permission}

\item{pop_county}{data frame: population table for state}

\item{sale}{data frame: sales table for permission}

\item{dashboard_yrs}{numeric: years to be included in summary}

\item{part_ref}{list: summary of parent permission to be used as reference 
for privilege rate. If NULL, calculates participation rate using pop_county}

\item{res_type}{identifies residency specific permissions ("Resident", 
"Nonresident", NULL)}

\item{segs}{character: vector of segment names to summarize}

\item{tests}{test values to pass to participants & churn calculation}

\item{tests_recruits}{test values to pass to recruits calculation}

\item{scaleup_test}{test values to pass to scaleup_part()}

\item{rate_test}{test values to pass to est_rate()}
}
\description{
Calculate all 5 metrics for a permission
}
\seealso{
Other dashboard functions: \code{\link{calc_part}},
  \code{\link{calc_rate}}, \code{\link{est_month}},
  \code{\link{est_pop}}, \code{\link{est_rate}},
  \code{\link{format_metrics}},
  \code{\link{format_result}}, \code{\link{write_dash}}
}
\concept{dashboard functions}
