% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{calc_part}
\alias{calc_part}
\title{Calculate participants: return list of length 2 (participants, residents)}
\usage{
calc_part(
  history,
  segs,
  tests,
  scaleup_test,
  res_type = NULL,
  use_recruits = FALSE
)
}
\arguments{
\item{history}{data frame: input history table for permission}

\item{segs}{character: vector of segment names to summarize}

\item{tests}{test values to pass to participants & churn calculation}

\item{scaleup_test}{test values to pass to scaleup_part()}

\item{res_type}{identifies residency specific permissions ("Resident", 
"Nonresident", NULL)}

\item{use_recruits}{logical: if TRUE, filter the history table with 
history$R3 == "Recruit"}
}
\description{
Calculate participants: return list of length 2 (participants, residents)
}
\seealso{
Other dashboard functions: 
\code{\link{calc_metrics}()},
\code{\link{calc_part_rate}()},
\code{\link{calc_priv_rate}()},
\code{\link{est_month}()},
\code{\link{est_pop}()},
\code{\link{est_rate}()},
\code{\link{fill_missing_months}()},
\code{\link{format_metrics}()},
\code{\link{write_dash}()}
}
\concept{dashboard functions}
