% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{calc_part_rate}
\alias{calc_part_rate}
\title{Calculate participation rate}
\usage{
calc_part_rate(part, pop_county, rate_test, res_type)
}
\arguments{
\item{part}{named list holding resident participant summary for privilege}

\item{pop_county}{data frame: population table for state}

\item{rate_test}{test values to pass to est_rate()}

\item{res_type}{identifies residency specific permissions ("Resident", 
"Nonresident", NULL)}
}
\description{
Refers to the ratio of permission resident participants to the state 
population.
}
\seealso{
Other dashboard functions: \code{\link{calc_metrics}},
  \code{\link{calc_part}}, \code{\link{calc_priv_rate}},
  \code{\link{est_month}}, \code{\link{est_pop}},
  \code{\link{est_rate}}, \code{\link{format_metrics}},
  \code{\link{write_dash}}
}
\concept{dashboard functions}
