% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{calc_priv_rate}
\alias{calc_priv_rate}
\title{Calculate privilege rate}
\usage{
calc_priv_rate(part, part_ref, rate_test)
}
\arguments{
\item{part}{named list that holds participant summary for given selected
privilege/subtype permission}

\item{part_ref}{name list that holds reference participant summary}

\item{rate_test}{test values to pass to est_rate()}
}
\description{
A privilege rate refers to the ratio between a given privilege (or subtype)
permission and a reference overall permission.
}
\seealso{
Other dashboard functions: 
\code{\link{calc_metrics}()},
\code{\link{calc_part_rate}()},
\code{\link{calc_part}()},
\code{\link{est_month}()},
\code{\link{est_pop}()},
\code{\link{est_rate}()},
\code{\link{fill_missing_months}()},
\code{\link{format_metrics}()},
\code{\link{write_dash}()}
}
\concept{dashboard functions}
