% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{calc_rate}
\alias{calc_rate}
\title{Calculate participation rate (or privilege rate if part_ref is not NULL)}
\usage{
calc_rate(part_res, pop_county, part, part_ref = NULL, res_type = NULL,
  rate_test = 1)
}
\arguments{
\item{part_res}{(for part. rate) named list that holds participant summary}

\item{pop_county}{(for part. rate) data frame that holds population data}

\item{part}{(for priv. rate) named list that holds participant summary}

\item{part_ref}{list: summary of parent permission to be used as reference 
for privilege rate. If NULL, calculates participation rate using pop_county}

\item{res_type}{identifies residency specific permissions ("Resident", 
"Nonresident", NULL)}

\item{rate_test}{test values to pass to est_rate()}
}
\description{
Calculate participation rate (or privilege rate if part_ref is not NULL)
}
\seealso{
Other dashboard functions: \code{\link{calc_metrics}},
  \code{\link{calc_part}}, \code{\link{est_month}},
  \code{\link{est_pop}}, \code{\link{est_rate}},
  \code{\link{format_metrics}},
  \code{\link{format_result}}, \code{\link{write_dash}}
}
\concept{dashboard functions}
