% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{dashboard}
\alias{dashboard}
\title{Sample Data: Metrics formatted for Dashboard Input}
\format{A data frame with 7 variables
\describe{
\item{quarter}{Time period covered (2 or 4)}
\item{group}{Permission group ("all_sports", "hunt", "fish", etc.)}
\item{segment}{Breakout ("All", "Age", "Gender", "Residency", "County", "month")}
\item{year}{Calendar Year (2008, 2009, ..., 2019)}
\item{category}{Category breakouts for segment}
\item{metric}{Metric Summarized ("participants", "recruits", "churn", "rate")}
\item{value}{Value of metric for given dimension}
}}
\description{
This data frame holds national/regional dashboard metrics formatted for 
dashboard input pulled from SC 2019-Q2 in Oct, 2019.
}
\seealso{
Other Sample Data: \code{\link{county_map_us}},
  \code{\link{hist_samp}}
}
\concept{Sample Data}
\keyword{datasets}
