% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{drop_na_custid}
\alias{drop_na_custid}
\title{Drop rows with missing cust_id from sale table}
\usage{
drop_na_custid(sale)
}
\arguments{
\item{sale}{data frame: sale table}
}
\description{
Not sure why the sales table would contain missing customer IDs, but there
are usually a handfull of NAs.
}
\seealso{
Other functions for producing license history: \code{\link{load_license}},
  \code{\link{write_history}}
}
\concept{functions for producing license history}
