% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{est_month}
\alias{est_month}
\title{Estimate monthly sales}
\usage{
est_month(sale, history, dashboard_yrs, use_recruits = FALSE)
}
\arguments{
\item{sale}{data frame: table to use for summarizing sales by month}

\item{history}{data frame: history table used for filter if use_recruits = TRUE}

\item{dashboard_yrs}{numeric: years to be included in summary}

\item{use_recruits}{logical: if TRUE, only include records where 
history$R3 == "Recruit"}
}
\description{
Estimate monthly sales
}
\seealso{
Other dashboard functions: \code{\link{calc_metrics}},
  \code{\link{calc_part}}, \code{\link{calc_rate}},
  \code{\link{est_pop}}, \code{\link{est_rate}},
  \code{\link{format_metrics}}, \code{\link{write_dash}}
}
\concept{dashboard functions}
