% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{est_rate}
\alias{est_rate}
\title{Estimate rate based on participant and population counts}
\usage{
est_rate(part_estimate, pop_estimate, test_threshold = 0.5)
}
\arguments{
\item{part_estimate}{data frame: table that holds participant counts}

\item{pop_estimate}{data frame: table that holds population counts}

\item{test_threshold}{numeric: threshold for output rate (e.g., 0.5 = 50
percent). A warning will be printed for every row that exceeds the threshold}
}
\description{
Estimate rate based on participant and population counts
}
\seealso{
Other dashboard functions: 
\code{\link{calc_metrics}()},
\code{\link{calc_part_rate}()},
\code{\link{calc_part}()},
\code{\link{calc_priv_rate}()},
\code{\link{est_month}()},
\code{\link{est_pop}()},
\code{\link{format_metrics}()},
\code{\link{write_dash}()}
}
\concept{dashboard functions}
