% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{extrapolate_pop}
\alias{extrapolate_pop}
\title{Extrapolate population forward for years without estimates}
\usage{
extrapolate_pop(pop, yrs)
}
\arguments{
\item{pop}{data frame: input population data for state}

\item{yrs}{numeric: yaers to include in summary}
}
\description{
Extrapolate population forward for years without estimates
}
\seealso{
Other functions to prepare data for summarization: \code{\link{aggregate_pop}},
  \code{\link{prep_pop}}, \code{\link{quarterly_filter}},
  \code{\link{quarterly_lapse}},
  \code{\link{recode_history}}
}
\concept{functions to prepare data for summarization}
