% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{factor_age_acs}
\alias{factor_age_acs}
\title{Convert age_acs from character to factor}
\usage{
factor_age_acs(pop, suppress_check = TRUE)
}
\arguments{
\item{pop}{data frame: input population data for state}

\item{suppress_check}{logical: If TRUE, does not print a coding summary}
}
\description{
This will ensure correct ordering in plots, etc. Uses \code{\link{age_map}}
sample data to identify factor levels.
}
\examples{
# db_census <- "E:/SA/Data-production/Data-Dashboards/_Shared/census.sqlite3"
# pop <- load_pop(db_census, "NC")
# pop <- factor_age_acs(pop, suppress_check = FALSE)
}
\seealso{
Other functions to prepare data for summarization: 
\code{\link{aggregate_pop}()},
\code{\link{extrapolate_pop}()},
\code{\link{prep_pop}()},
\code{\link{quarterly_filter}()},
\code{\link{quarterly_lapse}()},
\code{\link{recode_history}()}
}
\concept{functions to prepare data for summarization}
