% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{fill_missing_months}
\alias{fill_missing_months}
\title{Set missing months to zero in month breakouts}
\usage{
fill_missing_months(dat)
}
\arguments{
\item{dat}{formatted data frame for tableau dashboard}
}
\description{
This takes the completed dashboard and ensures consistency in month ranges
shown across the different permissions
}
\seealso{
Other dashboard functions: 
\code{\link{calc_metrics}()},
\code{\link{calc_part_rate}()},
\code{\link{calc_part}()},
\code{\link{calc_priv_rate}()},
\code{\link{est_month}()},
\code{\link{est_pop}()},
\code{\link{est_rate}()},
\code{\link{format_metrics}()},
\code{\link{write_dash}()}
}
\concept{dashboard functions}
