% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{filter_demo}
\alias{filter_demo}
\title{Filter data for given variable}
\usage{
filter_demo(priv, var = "sex", var_select = c("Male", "Female"),
  var_all = c("Male", "Female"))
}
\arguments{
\item{priv}{data for one permission}

\item{var}{name of variable for filtering}

\item{var_select}{currently selted values of var}

\item{var_all}{set of values that var can take}
}
\description{
To be run from \code{\link{run_visual_dive}}. Applies a filter as needed
based on the checked categories.
}
\seealso{
Other data dive functions: \code{\link{load_cust_samp}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
