% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{format_metrics}
\alias{format_metrics}
\title{Format metrics (list) into a single table output (data frame)}
\usage{
format_metrics(metrics, timeframe, group = "hunt")
}
\arguments{
\item{metrics}{named list that holds summary data produced by calc_metrics()}

\item{timeframe}{character: value to store in the 'timeframe' variable of 
the output (e.g, 'full-year', 'mid-year')}

\item{group}{character: value to sore in the 'group' variable of the 
output (e.g., 'all_sports', 'fish', 'hunt')}
}
\description{
Format metrics (list) into a single table output (data frame)
}
\seealso{
Other dashboard functions: \code{\link{calc_metrics}},
  \code{\link{calc_part}}, \code{\link{calc_rate}},
  \code{\link{est_month}}, \code{\link{est_pop}},
  \code{\link{est_rate}}, \code{\link{format_result}},
  \code{\link{write_dash}}
}
\concept{dashboard functions}
