% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{format_result}
\alias{format_result}
\title{Slightly modified version of format_result() from salic}
\usage{
format_result(df, timeframe, group, rename_input = c("category", "year",
  "value"))
}
\arguments{
\item{df}{data frame: Input table (3 variables) with estimated metrics}

\item{timeframe}{character: value to store in the 'timeframe' variable of 
the output (e.g, 'full-year', 'mid-year')}

\item{group}{character: value to sore in the 'group' variable of the 
output (e.g., 'all_sports', 'fish', 'hunt')}

\item{rename_input}{character: generic names for input variables as they
will appear in the output}
}
\description{
Slightly modified version of format_result() from salic
}
\seealso{
Other dashboard functions: \code{\link{calc_metrics}},
  \code{\link{calc_part}}, \code{\link{calc_rate}},
  \code{\link{est_month}}, \code{\link{est_pop}},
  \code{\link{est_rate}}, \code{\link{format_metrics}},
  \code{\link{write_dash}}
}
\concept{dashboard functions}
