% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{get_county_map}
\alias{get_county_map}
\title{Load county spatial data}
\usage{
get_county_map(state)
}
\arguments{
\item{state}{abbreviation of state to pull}
}
\description{
This uses functions from 2 packages (maps, ggplot2) to pull
}
\examples{
library(ggplot2)
county_map <- get_county_map("SC")
ggplot(county_map) + 
    geom_polygon(aes(long, lat, group = county))
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{join_county_map}}, \code{\link{plot_county}},
  \code{\link{plot_month}}, \code{\link{plot_value2}},
  \code{\link{plotly_config}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
