% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{get_county_map_dive}
\alias{get_county_map_dive}
\title{Load county spatial data for data dive}
\usage{
get_county_map_dive(state)
}
\arguments{
\item{state}{abbreviation of state to pull}
}
\description{
Mostly a wrapper for \code{\link{get_county_map}} with some formatting used
in data dive.
}
\examples{
get_county_map_dive("WI")
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{load_cust_samp}},
  \code{\link{plot_county_dive}}, \code{\link{plot_dist}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
