% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{hist_samp}
\alias{hist_samp}
\title{Sample Data: Permission data for use in \code{\link{run_visual_dive}}}
\format{
a data frame with 9 variables and approximately 550K observations
\describe{
\item{priv}{name of permission}
\item{cust_id}{customer ID}
\item{year}{license year}
\item{lapse}{lapsed next year (1 = TRUE, 0 = FALSE)}
\item{R3}{R3 category}
\item{res}{state residency}
\item{sex}{gender}
\item{fips}{county fips code (without state portion of code)}
\item{age}{age category}
}
}
\description{
A 1 percent sample of one state's permission dynamics
}
\seealso{
Other Sample Data: 
\code{\link{age_map}},
\code{\link{county_map_us}},
\code{\link{dashboard}}
}
\concept{Sample Data}
\keyword{datasets}
