% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{int_breaks}
\alias{int_breaks}
\title{Use only integer values for a plot's axis labels}
\usage{
int_breaks(x, n = 5)
}
\arguments{
\item{x}{placeholder of vector to be plotted}

\item{n}{integer value to control}
}
\description{
A helper function to be included in \code{\link{plot_month}} and
\code{\link{plot_value2}} to ensure axis labels don't include non-integer
values.
}
\seealso{
Other functions to run dashboard visualization: \code{\link{get_county_map}},
  \code{\link{join_county_map}}, \code{\link{plot_county}},
  \code{\link{plot_month}}, \code{\link{plot_value2}},
  \code{\link{plotly_config}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
