% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{join_county_map}
\alias{join_county_map}
\title{Join dashboard with county spatial data}
\usage{
join_county_map(dashboard, county_map, county_census)
}
\arguments{
\item{dashboard}{summary \code{\link{dashboard}} data}

\item{county_map}{data produced by \code{\link{get_county_map}}}

\item{county_census}{county names by fips, to provided more precise joining
between dashboard results and county_sf shapefile}
}
\description{
This takes the ouptut of  \code{\link{get_county_map}} and joins with 
\code{\link{dashboard}} data. The result is a list split by segment. The 
county_census table is used for linking on a more precise variable (county_fips
as oppossed to county name).
}
\examples{
library(dplyr)
data(dashboard)

county_map <- get_county_map("SC")
county_census <- load_counties(state = "SC")
dash_list <- join_county_map(dashboard, county_map, county_census)

# produce a warning by using the wrong state
county_map <- get_county_map("ME")
county_census <- load_counties(state = "ME")
dash_list <- join_county_map(dashboard, county_map, county_census)

# Maine and South Carolina actually share one county name
x <- filter(dash_list$county, group == "all_sports", quarter == 4)
plot_county(x) \%>\% gridExtra::grid.arrange(grobs = .)
}
\seealso{
Other functions to run dashboard visualization: \code{\link{get_county_map}},
  \code{\link{int_breaks}}, \code{\link{plot_county}},
  \code{\link{plot_month}}, \code{\link{plot_value2}},
  \code{\link{plotly_config}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
