% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{join_county_sf}
\alias{join_county_sf}
\title{Join dashboard with county spatial data}
\usage{
join_county_sf(dashboard, county_sf)
}
\arguments{
\item{dashboard}{summary \code{\link{dashboard}} data}

\item{county_sf}{data produced by \code{\link{pull_county_sf}}}
}
\description{
This takes the ouptut of  \code{\link{pull_county_sf}} and joins with 
\code{\link{dashboard}} data. The result is a list with county results
separated from all other results
}
\examples{
library(dplyr)
data(dashboard)
county_sf <- pull_county_sf("SC")
df <- join_county_sf(dashboard, county_sf)

# produce a warning by using the wrong state
county_sf <- pull_county_sf("ME")
df <- join_county_sf(dashboard, county_sf)

# Maine and South Carolina actually share one county name
df$county <- filter(df$county, group == "all_sports", quarter == 4)
plot_county(df$county)
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{plot_county}}, \code{\link{plot_month}},
  \code{\link{plot_value2}}, \code{\link{pull_county_sf}},
  \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
