% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{load_cust_samp}
\alias{load_cust_samp}
\title{Load a 10 percent sample all sportspersons}
\usage{
load_cust_samp(db, yrs, pct = 10, group = "all_sports")
}
\arguments{
\item{db}{file path to sqlite database}

\item{yrs}{years to be included from sqlite table}

\item{pct}{Sample size to draw, in whole percentage points (defaults to 10
percent)}

\item{group}{name of selected permission}
}
\description{
Every customer who holds a hunting or fishing permission at some point over
the timeframe has an equal change of being selected.
}
\examples{
\dontrun{
db_history <- "E:/SA/Data-production/Data-Dashboards/IA/history.sqlite3"
cust_samp <- load_cust_samp(db_history, 2006:2018)
dplyr::glimpse(cust_samp)
}
}
\seealso{
\code{\link{new_dashboard}}

Other data dive functions: \code{\link{setup_data_dive}}
}
\concept{data dive functions}
