% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{load_cust_sample}
\alias{load_cust_sample}
\title{Load a 10 percent sample all sportspersons}
\usage{
load_cust_sample(db, yrs, pct = 10, group = "all_sports")
}
\arguments{
\item{db}{file path to sqlite database}

\item{yrs}{years to be included from sqlite table}

\item{pct}{Sample size to draw, in whole percentage point (defaults to 10
percent)}

\item{group}{name of selected permission}
}
\description{
Every customer who holds a hunting or fishing permission at some point over
the timeframe has an equal change of being selected.
}
\examples{
# include examps
}
\seealso{
\code{\link{new_dashboard}}

Other data dive functions: \code{\link{setup_data_dive}}
}
\concept{data dive functions}
