% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{load_history_10pct}
\alias{load_history_10pct}
\title{Load history data for a 10 percent customer sample}
\usage{
load_history_10pct(db, group)
}
\arguments{
\item{db}{file path to sqlite database}

\item{group}{name of selected permission}
}
\description{
Pulls in all history data for a particular permission (group) for a 
randomly-selected 10 percent sample of customers.
}
\examples{
# include examps
}
\seealso{
\code{\link{load_history}}

Other data dive functions: \code{\link{setup_data_dive}}
}
\concept{data dive functions}
