% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{load_license}
\alias{load_license}
\title{Load license data (cust, lic, sale) into a list}
\usage{
load_license(db_license, yrs)
}
\arguments{
\item{db_license}{file path to license.sqlite3}

\item{yrs}{years to include in license history}
}
\description{
Load license data (cust, lic, sale) into a list
}
\seealso{
Other functions for producing license history: \code{\link{drop_na_custid}},
  \code{\link{write_history}}
}
\concept{functions for producing license history}
