% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-load.R
\name{load_sqlite}
\alias{load_sqlite}
\alias{load_cust}
\alias{load_sale}
\alias{load_counties}
\alias{load_pop}
\alias{load_lic_ids}
\alias{load_history}
\title{Load sqlite data}
\usage{
load_sqlite(db, query)

load_cust(db)

load_sale(db, yrs)

load_counties(db, state)

load_pop(db, state)

load_lic_ids(db, group)

load_history(db, group, yrs)
}
\arguments{
\item{db}{file path to sqlite database}

\item{query}{DBI statement for sqlite connection}

\item{yrs}{years to be included from sqlite table}

\item{state}{2-character abbreviation of selected state}

\item{group}{name of selected permission}
}
\description{
Load sqlite data
}
\section{Functions}{
\itemize{
\item \code{load_cust}: Load customer data

\item \code{load_sale}: Load sales for selected years

\item \code{load_counties}: Load county fips-names for attaching to cust & pop_county

\item \code{load_pop}: Load population for selected state (by county-age-sex)

\item \code{load_lic_ids}: Load license IDs for selected permission

\item \code{load_history}: Load license history for permission & join customers
}}

