% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{new_dashboard}
\alias{new_dashboard}
\title{Setup a new state dashboard with default directories and template scripts}
\usage{
new_dashboard(
  state,
  period,
  package = "sadash",
  print_message = "A new dashboard project has been initialized",
  ...
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{package}{character: name of package where the template files live}

\item{print_message}{character: Message to print on success}

\item{...}{other arguments passed to \code{\link[lictemplate]{new_project}}}
}
\description{
This is intended to be run before data processing for a new state begins. 
It's just a wrapper for \code{\link[lictemplate]{new_project}} from the 
lictemplate package.
}
\examples{
# new_dashboard("YY", "2018-q4")
}
\seealso{
Other functions for making template files/folders: 
\code{\link{setup_dashboard}()},
\code{\link{setup_data_dive}()},
\code{\link{update_dashboard}()}
}
\concept{functions for making template files/folders}
