% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{nonres_county_to_na}
\alias{nonres_county_to_na}
\title{Set county_fips to missing if not resident}
\usage{
nonres_county_to_na(x)
}
\arguments{
\item{x}{license history with county_fips & res variables}
}
\description{
Just a convenience function to ensure county_fips isn't populated where
res == 0 or is.na(res) b/c we wouldn't want these showing up in the data dive
}
\seealso{
Other data dive functions: \code{\link{load_cust_samp}},
  \code{\link{run_visual_dive}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
