% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{plot_county_dive}
\alias{plot_county_dive}
\title{Chloropleth county plot for data dive}
\usage{
plot_county_dive(priv, county_map, metric = "participants", pct = 10)
}
\arguments{
\item{priv}{data for one permission}

\item{county_map}{spatial table produced by \code{\link{get_county_map_dive}}}

\item{metric}{name of metric to summarize, either "participants" or "churn"}

\item{pct}{sample size (in whole percentage points) for hist_samp}
}
\description{
Chloropleth county plot for data dive
}
\examples{
library(dplyr)
data(hist_samp)
priv <- filter(hist_samp, priv == "all_sports", year == 2014)
county_map <- get_county_map_dive("WI")
plot_county_dive(priv, county_map, pct = 1)
plot_county_dive(priv, county_map, "churn")
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{get_county_map_dive}},
  \code{\link{load_cust_samp}}, \code{\link{plot_dist}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{set_other_county_na}},
  \code{\link{setup_data_dive}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
