% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot distribution for selected variable-year}
\usage{
plot_dist(priv, var = "sex")
}
\arguments{
\item{priv}{data for one permission}

\item{var}{name of variable to be plotted}
}
\description{
To be run from \code{\link{run_visual_dive}} to show a bar plot distribution.
}
\examples{
library(dplyr)
data(hist_samp)
priv <- filter(hist_samp, priv == "all_sports", year == 2015)
plot_dist(priv, "age")
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{load_cust_samp}}, \code{\link{plot_trend}},
  \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
