% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_pct}
\alias{plot_pct}
\title{Plot percent change by year for a given segment}
\usage{
plot_pct(df, seg, plot_title = "", measure = "pct_change",
  pct_range = 0.5)
}
\arguments{
\item{df}{data frame with summary results}

\item{seg}{variable name of segment to include in plot}

\item{plot_title}{caption to show in plot}

\item{measure}{variable to be plotted on the y axis}

\item{pct_range}{y-axis range}
}
\description{
Plot percent change by year for a given segment
}
\seealso{
Other functions to run dashboard visualization: \code{\link{plot_bar}},
  \code{\link{plot_segment}}, \code{\link{plot_value}}
}
\concept{functions to run dashboard visualization}
