% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_segment}
\alias{plot_segment}
\title{Run either plot_value() or plot_pct() based on measure}
\usage{
plot_segment(df, seg, plot_title = "", measure, pct_range)
}
\arguments{
\item{df}{data frame with summary results}

\item{seg}{variable name of segment to include in plot}

\item{plot_title}{caption to show in plot}

\item{measure}{variable to be plotted on the y axis}

\item{pct_range}{y-axis range}
}
\description{
Run either plot_value() or plot_pct() based on measure
}
\seealso{
Other functions to run dashboard visualization: \code{\link{plot_bar}},
  \code{\link{plot_pct}}, \code{\link{plot_value}}
}
\concept{functions to run dashboard visualization}
