% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot trendline for run_visual_dive()}
\usage{
plot_trend(tbl, down = "None", across = "None")
}
\arguments{
\item{tbl}{data frame holding table produced by \code{\link{summarize_trend}}}

\item{down}{variable name for optional facetting down. If "None", no facetting
will be done.}

\item{across}{variable name for optional facetting across. If "None", no facetting
will be done.}
}
\description{
To be run from \code{\link{run_visual_dive}}. 
This will plot a count of participants, optionally facetted down and/or across
by additional variables. If facetting (i.e., grouping) missing values will
be removed from the corresponding variables before plotting.
}
\examples{
library(dplyr)
data(hist_samp)
priv <- filter(hist_samp, priv == "all_sports")

summarize_trend(priv, pct = 1) \%>\% plot_trend()
summarize_trend(priv, "sex", pct = 1) \%>\% plot_trend("sex")
summarize_trend(priv, "sex", "age", pct = 1) \%>\% plot_trend("sex", "age")

summarize_trend(priv, pct = 1, metric = "churn") \%>\% plot_trend()
summarize_trend(priv, "sex", pct = 1, metric = "churn") \%>\% plot_trend("sex")
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{get_county_map_dive}},
  \code{\link{load_cust_samp}},
  \code{\link{plot_county_dive}}, \code{\link{plot_dist}},
  \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{set_other_county_na}},
  \code{\link{setup_data_dive}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
