% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_value}
\alias{plot_value}
\title{Plot value by year for a given segment}
\usage{
plot_value(df, seg, plot_title = "", measure = "value")
}
\arguments{
\item{df}{data frame with summary results}

\item{seg}{variable name of segment to include in plot}

\item{plot_title}{caption to show in plot}

\item{measure}{variable to be plotted on the y axis}
}
\description{
Plot value by year for a given segment
}
\seealso{
Other functions to run dashboard visualization: \code{\link{plot_bar}},
  \code{\link{plot_pct}}, \code{\link{plot_segment}}
}
\concept{functions to run dashboard visualization}
