% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_value2}
\alias{plot_value2}
\title{Plot value by year for segment}
\usage{
plot_value2(df, seg, plot_title = "", measure = "value", n = 5)
}
\arguments{
\item{df}{data frame with summary results}

\item{seg}{variable name of segment to include in plot}

\item{plot_title}{caption to show in plot}

\item{measure}{variable to be plotted on the y axis}

\item{n}{passed to \code{\link{int_breaks}} for x-axis labelling}
}
\description{
Mostly a wrapper for \code{\link[dashtemplate]{plot_bar}} with some 
additional formatting.
}
\examples{
library(dplyr)
data(dashboard)
filter(dashboard, group == "all_sports", quarter == 4) \%>\%
    plot_value2("Gender", "By Gender")
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{join_county_sf}}, \code{\link{plot_county}},
  \code{\link{plot_month}}, \code{\link{pull_county_sf}},
  \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
