% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{prep_history}
\alias{prep_history}
\title{Prepare license history for selected quarter (recoding)}
\usage{
prep_history(history, month_to_quarter = function(x) case_when(x <= 3 ~
  1, x <= 6 ~ 2, x <= 9 ~ 3, TRUE ~ 4))
}
\arguments{
\item{history}{data frame: table produced by load_history()}

\item{month_to_quarter}{function used to calculate quarter (based on month values)}
}
\description{
Prepare license history for selected quarter (recoding)
}
\seealso{
Other functions to prepare data for summarization: \code{\link{aggregate_pop}},
  \code{\link{extrapolate_pop}}, \code{\link{prep_pop}},
  \code{\link{quarterly_filter}},
  \code{\link{quarterly_lapse}}
}
\concept{functions to prepare data for summarization}
