% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{prep_pop}
\alias{prep_pop}
\title{Prepare population data for dashboard summaries - to be called after load_pop()}
\usage{
prep_pop(pop, yrs)
}
\arguments{
\item{pop}{data frame: input population data for state}

\item{yrs}{numeric: yaers to include in summary}
}
\description{
This is partly a wrapper for aggregate_pop() & extrapolate_pop()
}
\seealso{
Other functions to prepare data for summarization: \code{\link{aggregate_pop}},
  \code{\link{extrapolate_pop}},
  \code{\link{factor_age_acs}},
  \code{\link{quarterly_filter}},
  \code{\link{quarterly_lapse}},
  \code{\link{recode_history}}
}
\concept{functions to prepare data for summarization}
