% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{pull_county_sf}
\alias{pull_county_sf}
\title{Pull county spatial data}
\usage{
pull_county_sf(state)
}
\arguments{
\item{state}{abbreviation of state to pull}
}
\description{
This uses the urbnmapr package to pull a spatial features shapefile
}
\examples{
library(ggplot2)
county_sf <- pull_county_sf("SC")
ggplot(county_sf, aes()) + geom_sf(fill = "grey")
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{plot_county}}, \code{\link{plot_month}},
  \code{\link{plot_value2}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
