% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{quarterly_filter}
\alias{quarterly_filter}
\title{Filter by quarter in prepartion for calculating metrics}
\usage{
quarterly_filter(df, quarter, select_quarter, yrs)
}
\arguments{
\item{df}{data frame that holds data (e.g., sales or license history) for 
selected permission}

\item{quarter}{current quarter}

\item{select_quarter}{quarter to be summarized}

\item{yrs}{years to summarize}
}
\description{
There are 2 basic filter scenarios: (1) the select_quarter is less than 4
(i.e., partial year) and records for later quarters are removed, or (2)
the current quarter (quarter argument) is behind select_quarter, in which
case the most recent year is dropped.
}
\seealso{
Other functions to prepare data for summarization: \code{\link{aggregate_pop}},
  \code{\link{extrapolate_pop}}, \code{\link{prep_pop}},
  \code{\link{quarterly_lapse}},
  \code{\link{recode_history}}
}
\concept{functions to prepare data for summarization}
