% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{quarterly_lapse}
\alias{quarterly_lapse}
\title{Calculate lapse for a quarterly subset (where quarter != 4)}
\usage{
quarterly_lapse(history, select_quarter, yrs)
}
\arguments{
\item{history}{data frame that holds history table for selected permission}

\item{select_quarter}{quarter to be summarized}

\item{yrs}{years to summarize}
}
\description{
Calculate lapse for a quarterly subset (where quarter != 4)
}
\seealso{
Other functions to prepare data for summarization: 
\code{\link{aggregate_pop}()},
\code{\link{extrapolate_pop}()},
\code{\link{factor_age_acs}()},
\code{\link{prep_pop}()},
\code{\link{quarterly_filter}()},
\code{\link{recode_history}()}
}
\concept{functions to prepare data for summarization}
