% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-prep.R
\name{recode_history}
\alias{recode_history}
\title{Prepare license history for selected quarter (recoding)}
\usage{
recode_history(
  history,
  month_to_quarter = function(x) case_when(x <= 3 ~ 1, x <= 6 ~ 2, x <= 9 ~ 3, TRUE ~
    4)
)
}
\arguments{
\item{history}{data frame: table produced by load_history()}

\item{month_to_quarter}{function used to calculate quarter (based on month values).
The default function assumes a calendar year, so it will need to be modified
in cases where fiscal year is used (see VA 2019-q2 code for an example).}
}
\description{
This is basically a wrapper for several functions called in sequence, particularly
\code{\link[salic]{label_categories}} and \code{\link[salic]{recode_agecat}}.
The month_to_quarter parameter is used to specify an output "quarter" variable
based on the input "month" variable.
}
\seealso{
Other functions to prepare data for summarization: 
\code{\link{aggregate_pop}()},
\code{\link{extrapolate_pop}()},
\code{\link{factor_age_acs}()},
\code{\link{prep_pop}()},
\code{\link{quarterly_filter}()},
\code{\link{quarterly_lapse}()}
}
\concept{functions to prepare data for summarization}
