% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{remove_table}
\alias{remove_table}
\title{Convenience function to remove a table from a database}
\usage{
remove_table(db, table_name)
}
\arguments{
\item{db}{file path to database}

\item{table_name}{name of table to remove}
}
\description{
It can be helpful to make a fresh start with the permission table to avoid
potentially cryptic errors if column types change (e.g., lic_id stored 
as integer vs. character). This happened in 2019-q4 in WI.
}
\seealso{
Other functions for producing license history: 
\code{\link{data_check_sa}()},
\code{\link{drop_na_custid}()},
\code{\link{load_license}()},
\code{\link{write_history}()}
}
\concept{functions for producing license history}
