% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{run_visual}
\alias{run_visual}
\title{Run shiny app summary of dashboard results}
\usage{
run_visual(dashboard)
}
\arguments{
\item{dashboard}{a dataframe formatted as a \code{\link[salic]{dashboard}}
table described in salic}
}
\description{
Run shiny app summary of dashboard results
}
\examples{
data(dashboard)

\dontrun{
run_visual(dashboard)
}
}
\concept{functions to run dashboard visualization}
