% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{sapply2}
\alias{sapply2}
\title{Convenience function for use in calc_metrics() that wraps sapply}
\usage{
sapply2(x, ...)
}
\arguments{
\item{x}{expression for sapply}

\item{...}{other arguments passed to sapply}
}
\description{
Convenience function for use in calc_metrics() that wraps sapply
}
\keyword{internal}
