% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{set_nonres_county_na}
\alias{set_nonres_county_na}
\title{Set county_fips to missing if not resident}
\usage{
set_nonres_county_na(x)
}
\arguments{
\item{x}{license history with county_fips & res variables}
}
\description{
A convenience function to ensure county_fips isn't populated where
res == 0 or is.na(res). We wouldn't want these showing up in the data dive.
}
\seealso{
Other data dive functions: \code{\link{load_cust_samp}},
  \code{\link{run_visual_dive}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
