% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{set_other_county_na}
\alias{set_other_county_na}
\title{Set cust$county_fips to missing if from another state}
\usage{
set_other_county_na(cust, state, county_map = function()
  get_county_map_dive(state, FALSE))
}
\arguments{
\item{cust}{customer table}

\item{state}{2-character state abbreviation}

\item{county_map}{function to use for pulling a list of counties in state}
}
\description{
We may have counties from entirely different states attached to the customer
table and we don't want these in the data dive
}
\examples{
library(dplyr)

\dontrun{
db_history <- "E:/SA/Data-production/Data-Dashboards/IA/history.sqlite3"  
db_license <- "E:/SA/Data-production/Data-Dashboards/IA/license.sqlite3"
cust_samp <- left_join(
    load_cust_samp(db_history, 2006:2018),
    load_cust(db_license)
)
count(cust_samp, county_fips)
set_other_county_na(cust_samp, "IA") \%>\% count(county_fips)
}
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{get_county_map_dive}},
  \code{\link{load_cust_samp}},
  \code{\link{plot_county_dive}}, \code{\link{plot_dist}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
