% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{setup_dashboard}
\alias{setup_dashboard}
\title{Helper function for for dashboard templates}
\usage{
setup_dashboard(state, time_period, sa_path)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{time_period}{character: Time period for the first dashboard 
(e.g., "2015", "2016-q1", etc.)}

\item{sa_path}{character: File path to the Southwick main folder 
(for analysis and data, etc.)}
}
\description{
Called from \code{\link{new_dashboard}} and \code{\link{update_dashboard}}. 
Runs initial variable preparation, sets up folders, and creates a ".Rproj"
file in the analysis folder.
}
\seealso{
Other functions for making template files/folders: 
\code{\link{new_dashboard}()},
\code{\link{update_dashboard}()}
}
\concept{functions for making template files/folders}
