% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{summarize_trend}
\alias{summarize_trend}
\title{Summarize trend for either participants or churn}
\usage{
summarize_trend(priv, down = "None", across = "None",
  metric = "participants", pct = 10)
}
\arguments{
\item{priv}{data for one permission}

\item{down}{variable name for optional facetting down. If "None", no facetting
will be done.}

\item{across}{variable name for optional facetting across. If "None", no facetting
will be done.}

\item{metric}{name of metric to summarize, either "participants" or "churn"}

\item{pct}{sample size (in whole percentage points) for hist_samp}
}
\description{
To be run from \code{\link{run_visual_dive}} for preparing data for 
\code{\link{plot_trend}}
}
\examples{
data(hist_samp)
priv <- dplyr::filter(hist_samp, priv == "all_sports")
summarize_trend(priv, pct = 1)
summarize_trend(priv, "sex", "age", pct = 1)

summarize_trend(priv, metric = "churn")
summarize_trend(priv, "age", metric = "churn")
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{get_county_map_dive}},
  \code{\link{load_cust_samp}},
  \code{\link{plot_county_dive}}, \code{\link{plot_dist}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
