% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{ui_button_layout}
\alias{ui_button_layout}
\title{Define UI layout for run_visual()}
\usage{
ui_button_layout(dash_list)
}
\arguments{
\item{dash_list}{a list produced after running \code{\link{join_county_map}}}
}
\description{
A convenience function run from \code{\link{run_visual}} and
\code{\link{run_visual_county}}
}
\seealso{
Other functions to run dashboard visualization: \code{\link{get_county_map}},
  \code{\link{int_breaks}}, \code{\link{join_county_map}},
  \code{\link{plot_county}}, \code{\link{plot_month}},
  \code{\link{plot_value2}}, \code{\link{plotly_config}},
  \code{\link{run_visual_county}}
}
\concept{functions to run dashboard visualization}
