% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{update_dashboard}
\alias{update_dashboard}
\title{Setup a new period for a dashboard state}
\usage{
update_dashboard(
  state,
  time_period,
  ref_period,
  sa_path = "E:/SA",
  files_to_keep = c("\\\\.r", "documentation\\\\.tex")
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{time_period}{character: Time period for the first dashboard 
(e.g., "2015", "2016-q1", etc.)}

\item{ref_period}{folder name of reference time period (e.g., 2018-q4)}

\item{sa_path}{character: File path to the Southwick main folder 
(for analysis and data, etc.)}

\item{files_to_keep}{file matching for grep (not case sensitive). Files
with matching patterns will be copied from ref_period}
}
\description{
This will setup files based on a reference time period, rather than 
creating new files based on the default template. By default, the new analysis 
folder will include all ref_period R files (.R, .Rmd, etc.) and their
containing folders, as well as the documentation.tex file.
}
\seealso{
Other functions for making template files/folders: 
\code{\link{new_dashboard}()},
\code{\link{setup_dashboard}()},
\code{\link{setup_data_dive}()}
}
\concept{functions for making template files/folders}
