% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-metrics.R
\name{write_dash}
\alias{write_dash}
\title{Convenience function: output a csv file for given permission-quarter results}
\usage{
write_dash(dash, quarter, group, outdir = "code/3-dashboard-results/dash")
}
\arguments{
\item{dash}{data frame output of format_metrics()}

\item{quarter}{integer value of selected quarter}

\item{group}{character: value to sore in the 'group' variable of the 
output (e.g., 'all_sports', 'fish', 'hunt')}

\item{outdir}{file path for output directory}
}
\description{
Convenience function: output a csv file for given permission-quarter results
}
\seealso{
Other dashboard functions: 
\code{\link{calc_metrics}()},
\code{\link{calc_part_rate}()},
\code{\link{calc_part}()},
\code{\link{calc_priv_rate}()},
\code{\link{est_month}()},
\code{\link{est_pop}()},
\code{\link{est_rate}()},
\code{\link{format_metrics}()}
}
\concept{dashboard functions}
