% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{write_history}
\alias{write_history}
\title{Write history table to sqlite}
\usage{
write_history(history, group, lic_slct, db_history, db_license)
}
\arguments{
\item{history}{data frame: output history table}

\item{group}{name of permission group to output}

\item{lic_slct}{data frame: license table that identifies permission types}

\item{db_history}{sqlite history database}

\item{db_license}{sqlite license database}
}
\description{
Write history table to sqlite
}
\seealso{
Other functions for producing license history: 
\code{\link{data_check_sa}()},
\code{\link{drop_na_custid}()},
\code{\link{load_license}()},
\code{\link{remove_table}()}
}
\concept{functions for producing license history}
