% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_county}
\alias{plot_county}
\title{Make a county chloropleth}
\usage{
plot_county(df, measure, county_sf)
}
\arguments{
\item{df}{data frame with summary results}

\item{measure}{variable to be plotted on the y axis}

\item{county_sf}{spatial features dataset produced from \
code{\link{pull_county_sf}}}
}
\description{
Make a county chloropleth
}
\examples{
library(dplyr)
data(dashboard)
county_sf <- pull_county_sf("SC")
df <- filter(dashboard, group == "all_sports", quarter == 4)
plot_county(df, c("churn", "participants"), county_sf)
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{plot_month}}, \code{\link{plot_value2}},
  \code{\link{pull_county_sf}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
