% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_county}
\alias{plot_county}
\title{Make a county chloropleth for all metrics}
\usage{
plot_county(dat)
}
\arguments{
\item{dat}{a dashboard table where segment == "County" that has a "geometry"
column (e.g., produced from \code{\link{join_county_sf}})}
}
\description{
Intended to be run on the output of \code{\link{join_county_sf}} (the 
"county" element specifically). Uses \code{\link[gridExtra]{grid.arrange}}
to include a set of side-by-side chloropleths for every metric.
}
\examples{
library(dplyr)
data(dashboard)
county_sf <- pull_county_sf("SC")
county_census <- load_counties(state = "SC")
df <- join_county_sf(dashboard, county_sf, county_census)

df$county <- filter(df$county, group == "all_sports", quarter == 4)
plot_county(df$county)
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{join_county_sf}}, \code{\link{plot_month}},
  \code{\link{plot_value2}}, \code{\link{plotly_config}},
  \code{\link{pull_county_sf}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
