% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot trendline for run_visual_dive()}
\usage{
plot_trend(hist_samp, down = "None", across = "None", pct = 10)
}
\arguments{
\item{hist_samp}{data frame with a sample of license history for all privileges 
containing at least 9 variables: priv, cust_id, year, lapse, R3, res, sex, fips, age}

\item{down}{variable name for optional facetting down. If "None", no facetting
will be done.}

\item{across}{variable name for optional facetting across. If "None", no facetting
will be done.}

\item{pct}{sample size (in whole percentage points) for hist_samp}
}
\description{
To be run from \code{\link{run_visual_dive}}. 
This will plot a count of participants, optionally facetted down and/or across
by additional variables. If facetting (i.e., grouping) missing values will
be removed from the corresponding variables before plotting.
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{load_cust_samp}}, \code{\link{plot_dist}},
  \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
