% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_value2}
\alias{plot_value2}
\title{Plot value by year for metric-category}
\usage{
plot_value2(df, plot_title = "", measure = "value", n = 5)
}
\arguments{
\item{df}{data frame with summary results}

\item{plot_title}{caption to show in plot}

\item{measure}{variable to be plotted on the y axis}

\item{n}{passed to \code{\link{int_breaks}} for x-axis labelling}
}
\description{
Mostly a wrapper for \code{\link{plot_bar}} with some 
additional formatting. It's expected that the input table will only contain
a single group-quarter-segment.
}
\examples{
library(dplyr)
data(dashboard)
x <- filter(dashboard, group == "all_sports", quarter == 4, segment == "Gender")
plot_value2(x)
}
\seealso{
Other functions to run dashboard visualization: \code{\link{get_county_map}},
  \code{\link{int_breaks}}, \code{\link{join_county_map}},
  \code{\link{plot_bar}}, \code{\link{plot_county}},
  \code{\link{plot_month}}, \code{\link{plotly_config}},
  \code{\link{run_visual_county}},
  \code{\link{ui_button_layout}}
}
\concept{functions to run dashboard visualization}
