% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotly_config}
\alias{plotly_config}
\alias{plot_dash}
\alias{render_dash}
\title{Convenience functions for shiny plots}
\usage{
plotly_config(plot)

plot_dash(plot, height = "350px", ...)

render_dash(plot_code)
}
\arguments{
\item{plot}{Plot object for plotly}

\item{height}{passed to \code{\link[plotly]{plotlyOutput}}}

\item{...}{other arguments passed to \code{\link[plotly]{plotlyOutput}}}

\item{plot_code}{Code passed to \code{\link[plotly]{renderPlotly}}}
}
\description{
Use plotly_config() to remove extra plotly buttons. Use plot_dash() and
render_dash() for the shiny UI and Server respectively.
}
\seealso{
Other functions to run dashboard visualization: 
\code{\link{get_county_map}()},
\code{\link{int_breaks}()},
\code{\link{join_county_map}()},
\code{\link{plot_bar}()},
\code{\link{plot_county}()},
\code{\link{plot_month}()},
\code{\link{plot_value2}()},
\code{\link{run_visual_county}()},
\code{\link{ui_button_layout}()}
}
\concept{functions to run dashboard visualization}
