% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotly_config}
\alias{plotly_config}
\title{Define the modebar config for plotly}
\usage{
plotly_config(plot)
}
\arguments{
\item{plot}{Plot object for plotly}
}
\description{
Basically I want to remove the extra cruft at the top of the plots, except 
for the "save png" button.
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{join_county_sf}}, \code{\link{plot_county}},
  \code{\link{plot_month}}, \code{\link{plot_value2}},
  \code{\link{pull_county_sf}}, \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
