% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{run_visual}
\alias{run_visual}
\title{Run shiny app summary of dashboard results}
\usage{
run_visual(dash_list, include_county = FALSE)
}
\arguments{
\item{dash_list}{a list produced after running \code{\link{join_county_map}}}

\item{include_county}{if TRUE, county chloropleths will also be displayed}
}
\description{
The idea here is to replicate the functionality of the Tableau dashboard
to check/explore the results prior to sending to the Tableau analyst.
}
\examples{
data(dashboard)
county_map <- get_county_map("SC")
county_census <- load_counties(state = "SC")
dash_list <- join_county_map(dashboard, county_map, county_census)

\dontrun{
run_visual(dash_list)

# including county makes things a bit slow currently
run_visual(dash_list, include_county = TRUE)
}
}
\seealso{
Other functions to run dashboard visualization: \code{\link{get_county_map}},
  \code{\link{int_breaks}}, \code{\link{join_county_map}},
  \code{\link{plot_county}}, \code{\link{plot_month}},
  \code{\link{plot_value2}}, \code{\link{plotly_config}}
}
\concept{functions to run dashboard visualization}
