% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{run_visual}
\alias{run_visual}
\title{Run shiny app summary of dashboard results}
\usage{
run_visual(indir = "out", pct_range = 0.5)
}
\arguments{
\item{indir}{folder that holds summary results (in csv files)}

\item{pct_range}{y-axis range for percent change per year}
}
\description{
Run shiny app summary of dashboard results
}
