% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{run_visual}
\alias{run_visual}
\title{Run shiny app summary of dashboard results}
\usage{
run_visual(dashboard, state = "SC")
}
\arguments{
\item{dashboard}{a dataframe formatted as a \code{\link{dashboard}} table}

\item{state}{abbreviation of state to pull}
}
\description{
The idea here is to replicate the functionality of the Tableau dashboard
to check/explore the results prior to sending to the Tableau analyst.
}
\examples{
data(dashboard)

\dontrun{
run_visual(dashboard)
}
}
\seealso{
Other functions to run dashboard visualization: \code{\link{int_breaks}},
  \code{\link{plot_county}}, \code{\link{plot_month}},
  \code{\link{plot_value2}}, \code{\link{pull_county_sf}}
}
\concept{functions to run dashboard visualization}
