% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{run_visual_dive}
\alias{run_visual_dive}
\title{Run shiny app version of data dive}
\usage{
run_visual_dive(hist_samp, pct = 10)
}
\arguments{
\item{hist_samp}{data frame with a sample of license history for all privileges 
containing at least 9 variables: priv, cust_id, year, lapse, R3, res, sex, fips, age}

\item{pct}{sample size (in whole percentage points) for hist_samp}
}
\description{
This is a rough mock-up, intended to ensure (1) no surprises on the Tableau
end and (2) results look correct.
}
\examples{
\dontrun{
f <- "E:/SA/Data-production/Data-Dashboards/WI/2015-q4/WI-data-dive-10pct-2015q4/priv-WI-10pct.csv"
hist_samp <- readr::read_csv(f, progress = FALSE)
hist_samp <- salic::label_categories(hist_samp)
hist_samp <- salic::df_factor_age(hist_samp)
run_visual_dive(hist_samp)
}
}
\seealso{
\code{\link{run_visual}}

Other data dive functions: \code{\link{filter_demo}},
  \code{\link{load_cust_samp}}, \code{\link{plot_trend}},
  \code{\link{set_nonres_county_na}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
