% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{set_nonres_county_na}
\alias{set_nonres_county_na}
\title{Set county_fips to missing if not resident}
\usage{
set_nonres_county_na(x)
}
\arguments{
\item{x}{license history with county_fips & res variables}
}
\description{
Ensure county_fips isn't populated where res == 0 or is.na(res). There are
typically a few of these and we don't want them showing up in the data dive.
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{load_cust_samp}}, \code{\link{plot_trend}},
  \code{\link{run_visual_dive}},
  \code{\link{setup_data_dive}}
}
\concept{data dive functions}
