% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{set_nonres_county_na}
\alias{set_nonres_county_na}
\title{Set county_fips to missing if nonresident}
\usage{
set_nonres_county_na(x)
}
\arguments{
\item{x}{license history with county_fips & res variables}
}
\description{
Ensure county_fips isn't populated where res == 0 or is.na(res). There are
typically a few of these and we don't want them showing up in the data dive.
}
\seealso{
Other data dive functions: \code{\link{filter_demo}},
  \code{\link{get_county_map_dive}},
  \code{\link{load_cust_samp}},
  \code{\link{plot_county_dive}}, \code{\link{plot_dist}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_other_county_na}},
  \code{\link{setup_data_dive}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
