% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{setup_data_dive}
\alias{setup_data_dive}
\title{Setup data dive template}
\usage{
setup_data_dive(
  state = NULL,
  time_period = NULL,
  analysis_dir = getwd(),
  sa_path = "E:/SA",
  dive_dir = "5-data-dive"
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{time_period}{character: Time period for the first dashboard 
(e.g., "2015", "2016-q1", etc.)}

\item{analysis_dir}{full path name for analysis directory (only used if state
and time_period are NULL)}

\item{sa_path}{character: File path to the Southwick main folder 
(for analysis and data, etc.)}

\item{dive_dir}{folder for data dive code}
}
\description{
This is to be run for an existing state-time_period dashboard project. 
It creates a new folder with template code. Defaults to placing a "5-data-dive"
folder in the working directory (i.e., when state and time_period are NULL).
}
\examples{
# setup_data_dive()
}
\seealso{
\code{\link{new_dashboard}}

Other data dive functions: 
\code{\link{drop_state_code}()},
\code{\link{filter_demo}()},
\code{\link{get_county_map_dive}()},
\code{\link{load_cust_samp}()},
\code{\link{plot_county_dive}()},
\code{\link{plot_dist}()},
\code{\link{plot_trend}()},
\code{\link{run_visual_dive}()},
\code{\link{set_nonres_county_na}()},
\code{\link{set_other_county_na}()},
\code{\link{summarize_trend}()}
}
\concept{data dive functions}
