% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dive.R
\name{setup_data_dive}
\alias{setup_data_dive}
\title{Setup data dive template}
\usage{
setup_data_dive(state = NULL, time_period = NULL)
}
\arguments{
\item{state}{character: Two letter state designation. If NULL, uses the 
working directory}

\item{time_period}{character: Most recent time period ("2018-q4", etc.). If
NULL, uses the working directory}
}
\description{
This is to be run for an existing state-time_period dashboard project. 
It creates a "5-data-dive" folder with template code.
}
\examples{
# include examps
}
\seealso{
\code{\link{new_dashboard}}

Other data dive functions: \code{\link{filter_demo}},
  \code{\link{get_county_map_dive}},
  \code{\link{load_cust_samp}},
  \code{\link{plot_county_dive}}, \code{\link{plot_dist}},
  \code{\link{plot_trend}}, \code{\link{run_visual_dive}},
  \code{\link{set_nonres_county_na}},
  \code{\link{set_other_county_na}},
  \code{\link{summarize_trend}}
}
\concept{data dive functions}
