% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{update_dashboard}
\alias{update_dashboard}
\title{Setup a new period for a dashboard state}
\usage{
update_dashboard(state, time_period, ref_period, sa_path = "E:/SA")
}
\arguments{
\item{state}{character: Two letter state designation}

\item{time_period}{character: Time period for the first dashboard 
(e.g., "2015", "2016-q1", etc.)}

\item{ref_period}{folder name of reference time period (e.g., 2018-q4)}

\item{sa_path}{character: File path to the Southwick main folder 
(for analysis and data, etc.)}
}
\description{
This will setup files based on a reference time period, rather than 
creating new files based on the default template. The new analysis folder
will include all reference ".R" files and their containing folders.
}
\seealso{
Other functions for making template files/folders: \code{\link{new_dashboard}},
  \code{\link{setup_dashboard}}
}
\concept{functions for making template files/folders}
